reapi: Rename resource/hlapi to resource/reapi
Problem: In the current codebase, we mostly use the term reapi,
except for the name of the directory hlapi under resource. This
naming scheme should be consistent.

Rename resource/hlapi to resource/reapi and update relevant
makefiles and headers.
