+++ resource/reapi/bindings/c/Makefile.am
@@ -17,13 +17,13 @@ noinst_LTLIBRARIES = libreapi_cli.la libreapi_module.la
 libreapi_cli_la_SOURCES = \
     reapi_cli.cpp \
     reapi_cli.h \
-    $(top_srcdir)/resource/hlapi/bindings/c++/reapi.hpp \
-    $(top_srcdir)/resource/hlapi/bindings/c++/reapi_cli.hpp \
-    $(top_srcdir)/resource/hlapi/bindings/c++/reapi_cli_impl.hpp
+    $(top_srcdir)/resource/reapi/bindings/c++/reapi.hpp \
+    $(top_srcdir)/resource/reapi/bindings/c++/reapi_cli.hpp \
+    $(top_srcdir)/resource/reapi/bindings/c++/reapi_cli_impl.hpp
 
 libreapi_module_la_SOURCES = \
     reapi_module.cpp \
     reapi_module.h \
-    $(top_srcdir)/resource/hlapi/bindings/c++/reapi.hpp \
-    $(top_srcdir)/resource/hlapi/bindings/c++/reapi_module.hpp \
-    $(top_srcdir)/resource/hlapi/bindings/c++/reapi_module_impl.hpp
+    $(top_srcdir)/resource/reapi/bindings/c++/reapi.hpp \
+    $(top_srcdir)/resource/reapi/bindings/c++/reapi_module.hpp \
+    $(top_srcdir)/resource/reapi/bindings/c++/reapi_module_impl.hpp
